API Interna/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
oracledb==1.4.2
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.6
email-validator==2.1.0.post1
pydantic==2.4.2
pydantic-core==2.10.1 